Adding refactored mobile device modules with some error checking additions
